+++ packages/dara-core/js/index.tsx
@@ -96,6 +96,7 @@ export {
     DynamicComponent,
     Center,
     useAction,
+    useActionIsLoading,
     useVariable,
     getIcon,
     useDataVariable,