+++ packages/dara-components/changelog.md
@@ -6,6 +6,8 @@ title: Changelog
 
 -   Fixed an issue with `Card` where its content would render with zero height in some scenarios due to a missing `flex-grow` property.
 -   Clarified docs for `CausalGraphViewer`'s `editor_mode` prop.
+-   Fixed an issue where the `CausalGraphViewer` would error if the graph was empty in some cases
+-   Added a new 'Save as Image' button to the `CausalGraphViewer` UI to download the currently displayed graph pane as a high-resolution PNG image.
 
 ## 1.8.6